refactor: separate types into logical files
