+++ packages/app-headless-cms/src/types/shared.ts
@@ -1,0 +1,5 @@
+export interface CmsCreatedBy {
+    id: string;
+    displayName: string;
+    type: string;
+}
